+++ psrc/t8dev/cli/t8dev/emulator/virtualt.py
@@ -1,13 +1,10 @@
 ' VirtualT TRS-80 Model 100 (and friends) emulator suite. '
 
-from    shutil  import  copyfile, get_terminal_size
-from    urllib.request  import HTTPError
-import  textwrap
+from    contextlib  import chdir
+from    shutil  import  copy, copytree
 
-from    binary.romimage  import RomImage
 from    t8dev.cli.t8dev.emulator.suite  import Suite
 from    t8dev.cli.t8dev.util  import runtool
-import  os
 import  t8dev.path  as path
 
 class VirtualT(Suite):
@@ -15,5 +12,11 @@ class VirtualT(Suite):
     suitedesc = 'VirtualT TRS-80 Model 100 (and friends)'
 
     @classmethod
-    def setargparser(cls, parser):
-        return
+    def setargparser(cls, parser):  return
+
+    def run(self):
+        copytree(path.tool('src/virtualt/ROMs/'),
+                 self.emudir('ROMs'),
+                 dirs_exist_ok=True)
+        with chdir(self.emudir()):
+            runtool('virtualt')